apps/payments/serializers.py
from rest_framework import serializers
from apps.payments.models import Transaction

class TransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transaction
        fields = ('__all__')
