requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==5.0.7
django-allauth==0.63.3
django-filter==24.3
django-phone-verify==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
multidict==6.0.5
nexmo==2.5.2
oauthlib==3.2.2
packaging==24.1
phonenumbers==8.13.40
pillow==10.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.1.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
twilio==9.2.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
